requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.17.6
pandas==1.5.3
geopandas==1.0.1
numpy==1.26.4
annotated-types==0.7.0
pydantic==2.9.2
pydantic-core==2.23.4 
pydantic-settings==2.5.2 
python-dotenv==1.0.1